+++ client/app/components/EstimationMatrix/EstimationMatrix.js
@@ -11,6 +11,7 @@ import EstimationMatrixColumn from './EstimationMatrixColumn';
 
 import {StyledEMColumnsContainer, StyledEstimationMatrix, StyledNoStoriesHint} from './_styled';
 import {StyledStoryTitle} from '../_styled';
+import {getSettlingStories} from '../../state/commandTracking/commandTrackingSelectors';
 
 /**
  * Displays a table with all estimated stories (all stories with consensus), ordered by estimation value
@@ -18,6 +19,7 @@ import {StyledStoryTitle} from '../_styled';
 const EstimationMatrix = ({
   estimatedStories,
   cardConfig,
+  settlingStories,
   includeTrashedStories,
   toggleMatrixIncludeTrashed,
   settleEstimation
@@ -27,8 +29,10 @@ const EstimationMatrix = ({
   const [groupedStories, setGroupedStories] = useState({});
 
   useEffect(() => {
-    setGroupedStories(deriveGroupedStories(estimatedStories, includeTrashedStories));
-  }, [estimatedStories, includeTrashedStories]);
+    setGroupedStories(
+      deriveGroupedStories(estimatedStories, settlingStories, includeTrashedStories)
+    );
+  }, [estimatedStories, includeTrashedStories, settlingStories]);
 
   return (
     <StyledEstimationMatrix data-testid="matrix">
@@ -64,6 +68,7 @@ const EstimationMatrix = ({
 EstimationMatrix.propTypes = {
   estimatedStories: PropTypes.array,
   cardConfig: PropTypes.array,
+  settlingStories: PropTypes.array,
   includeTrashedStories: PropTypes.bool,
   toggleMatrixIncludeTrashed: PropTypes.func.isRequired,
   settleEstimation: PropTypes.func.isRequired
@@ -73,7 +78,8 @@ export default connect(
   (state) => ({
     estimatedStories: getAllStoriesWithConsensus(state),
     cardConfig: getCardConfigInOrder(state),
-    includeTrashedStories: state.ui.matrixIncludeTrashedStories
+    includeTrashedStories: state.ui.matrixIncludeTrashedStories,
+    settlingStories: getSettlingStories(state)
   }),
   {toggleMatrixIncludeTrashed, settleEstimation}
 )(EstimationMatrix);
@@ -87,19 +93,22 @@ const getColWidth = (cardConfigCount) => {
  * group stories according to estimation value.
  *
  * @param estimatedStories
+ * @param settlingStories
  * @param includeTrashedStories
  * @return {object} Object that maps estimation value (consensus) to array of stories
  */
-function deriveGroupedStories(estimatedStories, includeTrashedStories) {
+function deriveGroupedStories(estimatedStories, settlingStories, includeTrashedStories) {
   const stories = !includeTrashedStories
     ? estimatedStories.filter((s) => !s.trashed)
     : estimatedStories;
 
   return stories.reduce((total, current) => {
-    if (!total[current.consensus]) {
-      total[current.consensus] = [];
+    const isSettling = settlingStories.find((ss) => ss.storyId === current.id);
+    const consensus = isSettling ? isSettling.value : current.consensus;
+    if (!total[consensus]) {
+      total[consensus] = [];
     }
-    total[current.consensus].push(current);
+    total[consensus].push(current);
     return total;
   }, {});
 }